// src/types/course.tsx
export interface Course {
    id: string;
    title: string;
    detail: string;
    image: string;
    svg_icon: string;
    video_url: string;
    duration: number;
    duration_in: string;
    is_erased: boolean;
    course_type: string;
    created_by: string;
    updated_by: string;
    created_at: string;
    updated_at: string;
}